Add NamedAttachmentProxy specialization
